pyproject.toml
[project]
name = "trendsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.1",
]
